Applied CSS Fix to Sticky Footer - Courtesy of CSSTricks.com
I used the CSS Tricks Article by Chris Coyier with five ways of doing a sticky footer in order to apply changes to my sticky footer, which was having the problem of overlapping content. While doing so, I removed unreferenced items from my stylesheet so that it was easier for me to find the style items I am looking for as issues arise.